requirements.txt
av==8.0.2
discord.py==1.4.1
youtube-dl==2020.7.28